Bottles_BE/Bottles/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
Django==4.0.7
django-cors-headers==3.13.0
djangorestframework==3.14.0
frozenlist==1.4.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
multidict==6.0.4
mysqlclient==2.1.1
openai==0.28.1
Pillow==10.1.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.3.0
pytz==2022.2.1
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.3
tqdm==4.66.1
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2022.2
urllib3==2.0.7
yarl==1.9.2
zope.interface==6.1
